- /network: query gagal total ('ORDER BY position 0') utk user tanpa
  preferensi gender. BUG NYATA yg sedang terjadi di production.
- Pesan validasi kini terbaca user (hapus menu + jalur pembayaran kasir):
  pesan dari Error yang dilempar server action disensor Next.js di build
  produksi, jadi diubah ke pola return { ok, error }.
- Input voucher di panel Accept payment halaman detail order (sebelumnya
  hanya ada di panel kasir utama), khusus pelanggan terdaftar.
- Penanda pelanggan terdaftar vs tamu walk-in di halaman sesi (badge
  'guest') serta kartu meja kasir & waiter ('N registered').

Tanpa migrasi DB. Sudah dites di staging.

+++ src/lib/customer-actions.ts
@@ -846,9 +846,13 @@ export async function listAllMembers(opts?: {
     opts?.interestedIn === "male" || opts?.interestedIn === "female"
       ? opts.interestedIn
       : null;
+  // CATATAN: literal telanjang `0` TIDAK BOLEH dipakai di sini — di ORDER BY,
+  // Postgres membacanya sbg NOMOR KOLOM ("ORDER BY position 0 is not in select
+  // list") sehingga query gagal total saat tak ada preferensi gender.
+  // Di-cast supaya jadi ekspresi konstanta, bukan posisi kolom.
   const priorityExpr = wantGender
     ? sql<number>`CASE WHEN ${profiles.gender} = ${wantGender} THEN 0 ELSE 1 END`
-    : sql<number>`0`;
+    : sql<number>`0::int`;
   // Urutan level: VIP → Premium → Basic (rank efektif DESC). Dinegasikan
   // supaya tuple-compare keyset tetap ASC seragam.
   const tierExpr = sql<number>`-(${sqlEffectiveRank()})`;
